requirements.txt
ephem>=4.1.5
yfinance>=0.2.36
finnhub-python>=2.4.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
scikit-learn>=1.4.0
xgboost>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
